+++ CHANGELOG.md
@@ -38,11 +38,13 @@ and this project attempts to adhere to [Semantic Versioning](https://semver.org/
 -   Now using `nox` for testing.
 -   Build script has been moved from inside the package to the root of the repository.
 -   Added link to MIT license in all SVG files and split the license and Refactoring UI Inc. copyright information into separate lines.
+-   Icon registration has been simplified to just getting a list of the icon files under the `heroicons` template directory.
 
 ### Removed
 
 -   Dropped support for Python 3.7.
 -   Removed all build scripts and related methods from package.
+-   Removed `icons.py` and associated tests.
 
 ## [0.1.3]